feat(#84): [Tilak|Pooja] extract event mapper to separate class from rn event emitter
