+++ android/src/main/java/io/mosip/tuvali/rnModule/IRNEventEmitter.kt
@@ -1,7 +1,8 @@
 package io.mosip.tuvali.rnModule
 
+import com.facebook.react.bridge.WritableMap
 import io.mosip.tuvali.common.events.Event
 
 interface IRNEventEmitter {
-  fun emitEvent(event: Event)
+  fun emitEvent(eventMap: WritableMap)
 }
