+++ android/src/main/java/io/mosip/tuvali/rnModule/RNEventMapper.kt
@@ -1,0 +1,68 @@
+package io.mosip.tuvali.rnModule
+
+import com.facebook.react.bridge.Arguments
+import com.facebook.react.bridge.WritableMap
+import io.mosip.tuvali.common.events.*
+import kotlin.reflect.KVisibility
+import kotlin.reflect.full.memberProperties
+
+class RNEventMapper {
+  companion object {
+    fun toMap(event: Event): WritableMap {
+      val writableMap = Arguments.createMap()
+
+      writableMap.putString("type", getEventType(event))
+      populateProperties(event, writableMap)
+
+      return writableMap
+    }
+
+    private fun getEventType(event: Event): String {
+      return when (event) {
+        is DataReceivedEvent -> "onDataReceived"
+        is ErrorEvent -> "onError"
+        is VerificationStatusEvent -> "onVerificationStatusReceived"
+        is ConnectedEvent -> "onConnected"
+        is DataSentEvent -> "onDataSent"
+        is DisconnectedEvent -> "onDisconnected"
+        is SecureChannelEstablishedEvent -> "onSecureChannelEstablished"
+        else -> {
+          println("Invalid event type")
+          return ""
+        }
+      }
+    }
+
+    private fun populateProperties(event: Event, writableMap: WritableMap) {
+      event::class.memberProperties.forEach { property ->
+        if (property.visibility == KVisibility.PUBLIC) {
+          try {
+            val propertyValue = property.getter.call(event)
+            if (propertyValue is Enum<*>) {
+              handleEnumProperty(property.name, propertyValue, writableMap)
+            } else {
+              writableMap.putString(property.name, propertyValue.toString())
+            }
+          } catch (e: Exception) {
+            println("Unable to populate RN event ${property.name}")
+          }
+        }
+      }
+    }
+
+    private fun handleEnumProperty(propertyName: String, propertyValue: Enum<*>, writableMap: WritableMap) {
+      val valueField = propertyValue::class.memberProperties.firstOrNull { it.name == "value" }
+      when (val enumValue = valueField?.getter?.call(propertyValue)) {
+        is Int -> {
+          writableMap.putInt(propertyName, enumValue)
+        }
+        is String -> {
+          writableMap.putString(propertyName, enumValue)
+        }
+        else -> {
+          writableMap.putInt(propertyName, propertyValue.ordinal)
+        }
+      }
+    }
+  }
+}
