+++ ios/RNModules/RNEventEmitterProtocol.swift
@@ -1,5 +1,5 @@
 import Foundation
 
 protocol RNEventEmitterProtocol {
-    func emitEvent(event: Event)
+    func emitEvent(eventMap: NSMutableDictionary)
 }
